pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "LLM 기반 개인 지식 베이스 시스템"
requires-python = ">=3.11"
dependencies = [
    # 웹 클리핑
    "trafilatura>=1.12.0",
    "html2text>=2024.2.26",
    "requests>=2.31.0",
    # PDF 파싱
    "pymupdf>=1.24.0",
    # Office 파싱
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "python-pptx>=1.0.0",
    "python-docx>=1.1.0",
    # 이미지 처리
    "Pillow>=10.0.0",
    # LLM API
    "anthropic>=0.40.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    # 파일 감시
    "watchdog>=4.0.0",
    # 설정
    "pyyaml>=6.0.0",
]

[project.scripts]
kb = "scripts.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
